replace melt curve raw data builder with create method in quantstudio
